Reorganize events in Section 1
Events in Section 1 now occur in a more strict order, with the camera switching automatically as appropriate, removing the ability to manually switch between 1st-person view and orbiting camera view.

However, it is now possible to toggle the display inset on and off; in the latter case, the view intended to be rendered to the inset is now rendered to the main display (but with a bigger field of view).

Apart from the above, some additional cleanup is done in the intros to sections 2 and 3.